InsightX/requirements.txt
arxiv
python-dotenv
streamlit
llama-index
openai
PyPDF2